refactor: rename store function and place in service/share pkg
